some work on animations

+++ JonsEngine/include/Renderer/DirectX11/DX11Mesh.h
@@ -28,7 +28,7 @@ namespace JonsEngine
         DX11Mesh(ID3D11DevicePtr device, ID3D11DeviceContextPtr context, const std::vector<float>& vertexData, const std::vector<float>& normalData, const std::vector<float>& texCoords,
             const std::vector<float>& tangentData, const std::vector<uint16_t>& indexData, const Vec3& minBounds, const Vec3& maxBounds);
         DX11Mesh(ID3D11DevicePtr device, ID3D11DeviceContextPtr context, const std::vector<float>& vertexData, const std::vector<float>& normalData, const std::vector<float>& texCoords,
-            const std::vector<float>& tangentData, const std::vector<Mat4>& boneMatrices, const std::vector<uint16_t>& boneIndices, const std::vector<float>& boneWeights, const std::vector<uint16_t>& indexData, const Vec3& minBounds, const Vec3& maxBounds);
+            const std::vector<float>& tangentData, const std::vector<Mat4>& boneMatrices, const std::vector<uint8_t>& boneIndices, const std::vector<float>& boneWeights, const std::vector<uint16_t>& indexData, const Vec3& minBounds, const Vec3& maxBounds);
         ~DX11Mesh();
 
         void Draw();
