+++ JonsEngine/src/Renderer/DirectX11/DX11Renderer.cpp
@@ -23,7 +23,7 @@ namespace JonsEngine
     }
 
     DX11MeshID DX11Renderer::CreateMesh(const std::vector<float>& vertexData, const std::vector<float>& normalData, const std::vector<float>& texCoords, const std::vector<float>& tangentData,
-        const std::vector<Mat4>& bones, const std::vector<uint16_t>& boneIndices, const std::vector<float>& boneWeights, const std::vector<uint16_t>& indexData, const Vec3& minBounds, const Vec3& maxBounds)
+        const std::vector<Mat4>& bones, const std::vector<uint8_t>& boneIndices, const std::vector<float>& boneWeights, const std::vector<uint16_t>& indexData, const Vec3& minBounds, const Vec3& maxBounds)
     {
         return mImplementation->CreateMesh(vertexData, normalData, texCoords, tangentData, bones, boneIndices, boneWeights, indexData, minBounds, maxBounds);
     }
